pyproject.toml
[tool.poetry]
name = "photo-library-normalizer"
version = "0.1.0"
description = ""
authors = ["Jan Willhaus <mail@janwillhaus.de>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
piexif = "^1.1"
python-dateutil = "^2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
